Major refactoring, renaming and simplification.
