+++ modules/quant_purger/README.md
@@ -1,7 +1,25 @@
-# Quant cache tag purger
+# Quant Purger
 
-Adds a cache tag plugin which listens to Drupal invalidation events in order to
-queue Quant updates for related content.
+The Quant Purger helps you keep content fresh on your static Quant site
+after content updates within Drupal.
+
+## Purge Plugins
+
+This module is built on top of the [Purge module suite](https://www.drupal.org/project/purge).
+
+### Purger Plugin
+
+Processes cache invalidations based on type: 'everything', 'path' and 'tag'.
+The Quant cache will be purged based on these invalidations.
+
+- *Everything:* A site-wide cache purge, e.g. `/*`.
+- *Path:* Purges the given path.
+- *Tag:* Purges the given tag.
+
+### Queuer Plugin
+
+Adds a cache tag queuer plugin which listens to Drupal invalidation events in
+order to queue Quant updates for related content.
 
 For example, this allows node edits to trigger the main (`/node`) page to update
 along with any other pages associated with the node through cache tags (e.g.
@@ -13,6 +31,12 @@ several nodes, those nodes will be queued for updates when the term is edited.
 To ensure that queued content is processed in a timely manner, you can set up a
 Quant cron process that is separate from the core cron which just processes the
 Quant queue. This Quant cron can be run more regularly than the core cron.
+
+### TagsHeader Plugin
+
+Sets and formats the default response header with hashed cache tags.
+
+## Documentation
 
 See [Quant Purger documentation](https://docs.quantcdn.io/docs/integrations/drupal/purger)
 for additional information.
