+++ modules/quant_purger/config/schema/quant_purger.data_types.schema.yml
@@ -1,6 +1,6 @@
-quant_purge_header:
+quant_purger_header:
   type: mapping
-  label: 'Quant Header'
+  label: 'Quant Purger Header'
   mapping:
     field:
       type: string
