+++ modules/quant_purger/src/Plugin/Purge/Purger/QuantPurger.php
@@ -2,9 +2,13 @@
 
 namespace Drupal\quant_purger\Plugin\Purge\Purger;
 
-use Drupal\purge\Plugin\Purge\Purger\PurgerInterface;
+use Drupal\Core\Utility\Token;
 use Drupal\purge\Plugin\Purge\Invalidation\InvalidationInterface;
-use Drupal\quant_purger\Entity\Hash;
+use Drupal\purge\Plugin\Purge\Purger\PurgerBase;
+use Drupal\purge\Plugin\Purge\Purger\PurgerInterface;
+use Drupal\quant_purger\Entity\QuantPurgerSettings;
+use GuzzleHttp\ClientInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Quant Purger.
@@ -14,25 +18,115 @@ use Drupal\quant_purger\Entity\Hash;
  *   label = @Translation("Quant Purger"),
  *   configform = "\Drupal\quant_purger\Form\QuantPurgerConfigForm",
  *   cooldown_time = 0.2,
- *   description = @Translation("Purger that sends invalidation expressions from your Drupal instance to the QuantCDN platform."),
+ *   description = @Translation("Purger that sends invalidations from your Drupal site to the QuantCDN platform."),
  *   multi_instance = FALSE,
  *   types = {"everything", "path", "tag"},
  * )
  */
-class QuantPurger extends QuantPurgerBase implements PurgerInterface {
+class QuantPurger extends PurgerBase implements PurgerInterface {
+
+  /**
+   * The Guzzle HTTP client.
+   *
+   * @var \GuzzleHttp\Client
+   */
+  protected $client;
+
+  /**
+   * The settings entity holding all configuration.
+   *
+   * @var \Drupal\quant_purger\Entity\QuantPurgerSettings
+   */
+  protected $settings;
+
+  /**
+   * The token service.
+   *
+   * @var \Drupal\Core\Utility\Token
+   */
+  protected $token;
+
+  /**
+   * Constructs the Quant Purger.
+   *
+   * @param array $configuration
+   *   A configuration array containing information about the plugin instance.
+   * @param string $plugin_id
+   *   The plugin_id for the plugin instance.
+   * @param mixed $plugin_definition
+   *   The plugin implementation definition.
+   * @param \GuzzleHttp\ClientInterface $http_client
+   *   An HTTP client that can perform remote requests.
+   * @param \Drupal\Core\Utility\Token $token
+   *   The token service.
+   */
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, ClientInterface $http_client, Token $token) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->settings = QuantPurgerSettings::load($this->getId());
+    // Note: We use the Quant HTTP client rather than the generic Guzzle client.
+    $this->client = \Drupal::service('quant_api.client');
+    $this->token = $token;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('http_client'),
+      $container->get('token')
+    );
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getLabel() {
+    if ($this->settings->name) {
+      return $this->settings->name;
+    }
+    else {
+      return parent::getLabel();
+    }
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getTypes() {
+    // @todo Add url?
+    return [
+      'everything',
+      'path',
+      'tag',
+    ];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function delete() {
+    QuantPurgerSettings::load($this->getId())->delete();
+  }
 
   /**
    * {@inheritdoc}
    */
   public function invalidate(array $invalidations) {
 
+    // The processed data may include 'everything', 'paths' and/or 'tags'.
     $processed = $this->processInvalidations($invalidations);
 
-    if ($processed['everything']) {
+    // The 'everything' invaldiation type takes precedence.
+    if (!empty($processed['everything'])) {
       $this->invalidateEverything($invalidations);
       return;
     }
 
+    // Otherwise handle any paths and tags.
     if (!empty($processed['paths'])) {
       $this->invalidatePaths($processed['paths']);
     }
@@ -44,11 +138,57 @@ class QuantPurger extends QuantPurgerBase implements PurgerInterface {
   }
 
   /**
+   * Process invalidations based on type.
+   *
+   * @param array $invalidations
+   *   The array of Invalidation items to process.
+   *
+   * @return array
+   *   The processed array with 'everything', 'paths' and 'tags' keys.
+   */
+  public function processInvalidations(array $invalidations) {
+    $everything = FALSE;
+    $paths = [];
+    $tags = [];
+
+    foreach ($invalidations as $invalidation) {
+
+      if ($invalidation->getType() == 'everything') {
+        // 'Everything' takes precedence and will issue a site-wide purge.
+        $invalidation->setState(InvalidationInterface::PROCESSING);
+        $everything = TRUE;
+        $paths = [];
+        $tags = [];
+        break;
+      }
+      elseif ($invalidation->getType() == 'path') {
+        $invalidation->setState(InvalidationInterface::PROCESSING);
+        $paths[] = $invalidation;
+      }
+      elseif ($invalidation->getType() == 'tag') {
+        $invalidation->setState(InvalidationInterface::PROCESSING);
+        $tags[] = $invalidation;
+      }
+      else {
+        $invalidation->setState(InvalidationInterface::NOT_SUPPORTED);
+        $this->logger()->warning('Invalidation type not supported: @type', ['@type' => $invalidation->getType()]);
+      }
+    }
+
+    $processed = [
+      'everything' => $everything,
+      'paths' => $paths,
+      'tags' => $tags,
+    ];
+
+    return $processed;
+  }
+
+  /**
    * Invalidate with the path '/*' to purge the entire project cache.
    *
    * @param array $invalidations
-   *   This takes in an array of Invalidation objects, processing them all in a
-   *   loop, generally from the purge queue.
+   *   Array of Invalidation objects to process.
    */
   public function invalidateEverything(array $invalidations) {
 
@@ -60,7 +200,7 @@ class QuantPurger extends QuantPurgerBase implements PurgerInterface {
       }
     }
     catch (\Exception $e) {
-      $this->logger()->notice('Error attempting to purge entire cache: ' . $e->getMessage());
+      $this->logger()->error('Error attempting to purge entire cache: @message', ['@message' => $e->getMessage()]);
       error_log($e->getMessage());
       foreach ($invalidations as $invalidation) {
         $invalidation->setState(InvalidationInterface::FAILED);
@@ -79,12 +219,12 @@ class QuantPurger extends QuantPurgerBase implements PurgerInterface {
     foreach ($invalidations as $invalidation) {
       try {
         $path = '/' . $invalidation->getExpression();
-        $this->logger()->debug('[path] Purging path: ' . $path);
+        $this->logger()->debug('[path] Purging path: @path', ['@path' => $path]);
         $this->purgePath($path);
         $invalidation->setState(InvalidationInterface::SUCCEEDED);
       }
       catch (\Exception $e) {
-        $this->logger()->notice('Error attempting to purge cache path: ' . $e->getMessage());
+        $this->logger()->error('Error attempting to purge paths: @message', ['@message' => $e->getMessage()]);
         error_log($e->getMessage());
         $invalidation->setState(InvalidationInterface::FAILED);
       }
@@ -99,7 +239,8 @@ class QuantPurger extends QuantPurgerBase implements PurgerInterface {
    */
   public function invalidateTags(array $invalidations) {
     try {
-      $this->logger()->debug('[tags] Purging tags: ' . implode(' ', $invalidations));
+      // Log tags prior to hashing.
+      $this->logger()->debug('[tags] Purging tags: @tags', ['@tags' => implode(' ', $invalidations)]);
 
       $tags = [];
       foreach ($invalidations as $invalidation) {
@@ -110,10 +251,64 @@ class QuantPurger extends QuantPurgerBase implements PurgerInterface {
       $invalidation->setState(InvalidationInterface::SUCCEEDED);
     }
     catch (\Exception $e) {
-      $this->logger()->notice('Error attempting to purge cache path: ' . $e->getMessage());
+      $this->logger()->error('Error attempting to purge tags: @message', ['@message' => $e->getMessage()]);
       error_log($e->getMessage());
       $invalidation->setState(InvalidationInterface::FAILED);
     }
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getCooldownTime() {
+    return $this->settings->cooldown_time;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getIdealConditionsLimit() {
+    return $this->settings->max_requests;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getTimeHint() {
+    // When runtime measurement is enabled, we just use the base implementation.
+    if ($this->settings->runtime_measurement) {
+      return parent::getTimeHint();
+    }
+    // Theoretically connection timeouts and general timeouts can add up, so
+    // we add up our assumption of the worst possible time it takes as well.
+    return $this->settings->connect_timeout + $this->settings->timeout;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function hasRuntimeMeasurement() {
+    return (bool) $this->settings->runtime_measurement;
+  }
+
+  /**
+   * Sends a path-based purge request to the Quant API.
+   *
+   * @param string $path
+   *   The path to purge.
+   */
+  public function purgePath(string $path) {
+    $this->client->purgePath($path);
+  }
+
+  /**
+   * Sends a tags-based purge request to the Quant API.
+   *
+   * @param array $tags
+   *   The array of tags to purge.
+   */
+  public function purgeTags(array $tags) {
+    $this->client->purgeTags($tags);
   }
 
 }
