+++ modules/quant_purger/src/Plugin/Purge/Queuer/QuantPurgerQueuer.php
@@ -9,13 +9,12 @@ use Drupal\purge\Plugin\Purge\Queuer\QueuerInterface;
  * Quant Purger Queuer.
  *
  * @PurgeQueuer(
- *   id = "quant",
+ *   id = "quant_purger_queuer",
  *   label = @Translation("Quant Purger Queuer"),
- *   description = @Translation("Queue impacted content updates."),
+ *   description = @Translation("Update Quant queue based on content updates."),
  *   enable_by_default = true,
  *   types = {"tag"},
  *   configform = "\Drupal\quant_purger\Form\QuantPurgerQueuerConfigForm",
  * )
  */
-// @todo Change id to 'quant_purger_queuer' which requires an update hook.
 class QuantPurgerQueuer extends QueuerBase implements QueuerInterface {}
